move module-request to its own module
