+++ packages/core/src/index.ts
@@ -12,7 +12,8 @@ export { default as Options, optionsWithDefaults } from './options';
 export { default as WaitForTrees, OutputPaths } from './wait-for-trees';
 export { compile as jsHandlebarsCompile } from './js-handlebars';
 export { todo, unsupported, warn, debug, expectWarning, throwOnWarnings } from './messages';
-export { Resolver, ModuleRequest, Resolution } from './module-resolver';
+export { Resolver } from './module-resolver';
+export type { ModuleRequest, Resolution } from './module-request';
 export type { Options as ResolverOptions } from './module-resolver-options';
 export { ResolverLoader } from './resolver-loader';
 export { virtualContent } from './virtual-content';
